requirements.txt
# Core Flask packages
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Migrate==4.0.7

bcrypt==4.2.0
python-dotenv==1.0.1

requests==2.32.3
Werkzeug>=3.1.0
Jinja2==3.1.4

# Google Generative AI
google-generativeai==0.8.5
